+++ cranelift/codegen/src/isa/unwind.rs
@@ -3,14 +3,14 @@
 use serde::{Deserialize, Serialize};
 
 pub mod systemv;
+pub mod winx64;
 
 /// Represents unwind information for a single function.
 #[derive(Clone, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "enable-serde", derive(Serialize, Deserialize))]
 pub enum UnwindInfo {
     /// Windows x64 ABI unwind information.
-    #[cfg(feature = "x86")]
-    WindowsX64(super::x86::unwind::windows::UnwindInfo),
+    WindowsX64(winx64::UnwindInfo),
     /// System V ABI unwind information.
     SystemV(systemv::UnwindInfo),
 }
