+++ cranelift/codegen/src/isa/x86/abi.rs
@@ -952,7 +952,7 @@ pub fn create_unwind_info(
                 .map(|u| UnwindInfo::SystemV(u))
         }
         CallConv::WindowsFastcall => {
-            super::unwind::windows::create_unwind_info(func, isa, Some(RU::rbp.into()))?
+            super::unwind::winx64::create_unwind_info(func, isa, Some(RU::rbp.into()))?
                 .map(|u| UnwindInfo::WindowsX64(u))
         }
         _ => None,
