+++ cranelift/codegen/src/isa/x86/unwind.rs
@@ -1,4 +1,4 @@
 //! Module for x86 unwind generation for supported ABIs.
 
 pub mod systemv;
-pub mod windows;
+pub mod winx64;
